*Change: Adding resnet50_swish, modeyfication in layers/center_loss and modeling/baseline.py

+++ layers/center_loss.py
@@ -37,7 +37,7 @@ class CenterLoss(nn.Module):
         batch_size = x.size(0)
         distmat = torch.pow(x, 2).sum(dim=1, keepdim=True).expand(batch_size, self.num_classes) + \
                   torch.pow(self.centers, 2).sum(dim=1, keepdim=True).expand(self.num_classes, batch_size).t()
-        distmat.addmm_(1, -2, x, self.centers.t())
+        distmat.addmm_(1, -2, x.float(), self.centers.t().float())
 
         classes = torch.arange(self.num_classes).long()
         if self.use_gpu: classes = classes.cuda()
